requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
weaviate-client==4.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
